test(ToolsPage): fix test
