+++ packages/core/src/index.ts
@@ -44,4 +44,5 @@ export * from './typeschema/slices';
 export * from './typeschema/types';
 export * from './typeschema/validation';
 export * from './utils';
+export * from './version-utils';
 export * from './websockets/reconnecting-websocket';
